Move commitment and packet types
