+++ crates/chain/chain-components/src/traits/types/mod.rs
@@ -17,7 +17,6 @@ pub mod height;
 pub mod ibc;
 pub mod ibc_events;
 pub mod message;
-pub mod packet;
 pub mod packets;
 pub mod proof;
 pub mod status;
@@ -25,3 +24,4 @@ pub mod timestamp;
 pub mod update_client;
 
 pub use hermes_chain_type_components::traits::event;
+pub use hermes_chain_type_components::traits::ibc::packet;
