+++ crates/chain/chain-components/src/traits/types/proof.rs
@@ -2,10 +2,7 @@ use cgp::prelude::*;
 
 use crate::traits::types::height::HasHeightType;
 
-#[derive_component(CommitmentProofTypeComponent, ProvideCommitmentProofType<Chain>)]
-pub trait HasCommitmentProofType: Async {
-    type CommitmentProof: Async;
-}
+pub use hermes_chain_type_components::traits::commitment_proof::*;
 
 #[derive_component(CommitmentProofHeightGetterComponent, CommitmentProofHeightGetter<Chain>)]
 pub trait HasCommitmentProofHeight: HasCommitmentProofType + HasHeightType {
