+++ crates/chain/chain-type-components/src/traits/commitment_prefix.rs
@@ -1,0 +1,6 @@
+use cgp::prelude::*;
+
+#[derive_component(CommitmentPrefixTypeComponent, ProvideCommitmentPrefixType<Chain>)]
+pub trait HasCommitmentPrefixType: Async {
+    type CommitmentPrefix: Async;
+}
