+++ crates/chain/chain-type-components/src/traits/commitment_proof.rs
@@ -1,0 +1,6 @@
+use cgp::prelude::*;
+
+#[derive_component(CommitmentProofTypeComponent, ProvideCommitmentProofType<Chain>)]
+pub trait HasCommitmentProofType: Async {
+    type CommitmentProof: Async;
+}
