+++ crates/chain/chain-type-components/src/traits/mod.rs
@@ -1,3 +1,5 @@
+pub mod commitment_prefix;
+pub mod commitment_proof;
 pub mod event;
 pub mod height;
 pub mod ibc;
